Move PromisedValue to a separate file
